# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: JernihAI API
  description: 'API backend JernihAI — platform peningkatan kualitas gambar.


    Dokumentasi API Publik B2B (FR-14) untuk developer: `docs/API_B2B.md`. Spesifikasi
    OpenAPI lengkap: `docs/api/openapi.yaml` (regenerate via `python api/scripts/export_openapi.py`).'
  version: 0.1.0
paths:
...
  /api/v1/jobs:
    post:
      tags:
      - jobs
      summary: Upload gambar & mulai proses (FR-02/FR-03)
      operationId: create_job_api_v1_jobs_post
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_create_job_api_v1_jobs_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - jobs
      summary: Riwayat job user (FR-10)
      description: 'Daftar job milik user, terbaru dulu (FR-10).


        Hanya job milik user yang login — tidak membocorkan job orang lain.

        Pagination `limit`/`offset`; `result_deleted_at` memberitahu UI bahwa

        hasil sudah dihapus retensi (tombol unduh ulang dinonaktifkan).'
      operationId: list_jobs_api_v1_jobs_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobListOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/jobs/{job_id}:
    get:
      tags:
      - jobs
      summary: Cek status job (polling — FR-03)
      operationId: get_job_api_v1_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/admin/jobs:
    get:
      tags:
      - admin
      summary: Semua job lintas user (FR-13)
      description: 'Daftar job (terbaru dulu) — semua user atau riwayat SATU user
        (`email`).


        `email` = pencocokan persis (email unik) — dipakai admin melihat

        riwayat lengkap milik user tertentu.'
      operationId: admin_jobs_api_v1_admin_jobs_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: email
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Email
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminJobList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/users:
    get:
      tags:
      - admin
      summary: Daftar user + kuota & kredit (FR-13)
      description: 'Direktori user: email, nama, provider, kuota gratis (FR-06), saldo

        kredit (FR-11), consent privasi (FR-07), dan jumlah riwayat job.


        `email` = pencarian parsial (ilike); urut terbaru dulu.'
      operationId: admin_users_api_v1_admin_users_get
      parameters:
      - name: email
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Email
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/users/{user_id}:
    get:
      tags:
      - admin
      summary: Detail satu user (profil + kuota/kredit) — FR-13
      description: 'Profil lengkap satu user — dipakai halaman detail admin

        (email, kuota, kredit, consent, jumlah riwayat job).'
      operationId: admin_user_detail_api_v1_admin_users__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - admin
      summary: Aktif / nonaktifkan akun user (alat admin)
      description: 'Toggle `is_active` user (suspend). Sesi user langsung ditolak
        oleh

        `get_current_user`; login berikutnya juga ditolak. Admin tidak bisa

        menonaktifkan akun sendiri (mengunci diri).'
      operationId: admin_user_update_api_v1_admin_users__user_id__patch
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/users/{user_id}/transactions:
    get:
      tags:
      - admin
      summary: Transaksi kredit satu user — FR-13
      description: 'Riwayat pembelian kredit (FR-11) milik satu user — terbaru dulu.

        Dipakai halaman detail admin untuk audit transaksi.'
      operationId: admin_user_transactions_api_v1_admin_users__user_id__transactions_get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTransactionList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/quota/reset:
    post:
      tags:
      - admin
      summary: Reset kuota gratis user (alat admin)
      description: 'Reset pemakaian kuota gratis (FR-06) untuk satu user atau semua.


        Alat untuk pengelola/pengembang saat uji coba: set `used=0` dan tanggal

        reset ke hari ini (WIB) — tanpa mengubah limit dari env.'
      operationId: admin_quota_reset_api_v1_admin_quota_reset_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuotaResetRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotaResetOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/jobs/{job_id}:
    delete:
      tags:
      - admin
      summary: Hapus job + file-nya (alat admin)
      description: 'Hapus permanen job (baris DB + file original & hasil di storage).


        File hanya dihapus bila berada di dalam upload_dir/result_dir (guard

        path traversal via `delete_if_inside`; berlaku untuk backend lokal

        maupun R2). Dipakai pengelola membersihkan data uji coba.'
      operationId: admin_delete_job_api_v1_admin_jobs__job_id__delete
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Admin Delete Job Api V1 Admin Jobs  Job Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/export/users.csv:
    get:
      tags:
      - admin
      summary: 'Ekspor CSV: seluruh user + kuota/kredit (audit — FR-13)'
      description: 'CSV untuk audit: email, kuota, kredit, consent, jumlah riwayat.

        `email` = filter parsial (sama dengan direktori user).'
      operationId: admin_export_users_csv_api_v1_admin_export_users_csv_get
      parameters:
      - name: email
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Email
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/export/jobs.csv:
    get:
      tags:
      - admin
      summary: 'Ekspor CSV: seluruh job + email pemilik (audit — FR-13)'
      description: CSV riwayat job lintas user (atau satu user via `email` persis).
      operationId: admin_export_jobs_csv_api_v1_admin_export_jobs_csv_get
      parameters:
      - name: email
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Email
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_create_job_api_v1_jobs_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        scale:
          type: integer
          title: Scale
          default: 2
        output_format:
          type: string
          title: Output Format
          default: webp
        face_enhance:
          type: boolean
          title: Face Enhance
          default: false
        denoise:
          type: boolean
          title: Denoise
          default: false
        color_enhance:
          type: boolean
          title: Color Enhance
          default: false
      type: object
      required:
      - file
      title: Body_create_job_api_v1_jobs_post
...
    JobOut:
      properties:
        id:
          type: string
          title: Id
        status:
          type: string
          title: Status
        scale:
          type: integer
          title: Scale
        output_format:
          type: string
          title: Output Format
        face_enhance:
          type: boolean
          title: Face Enhance
          default: false
        denoise:
          type: boolean
          title: Denoise
          default: false
        color_enhance:
          type: boolean
          title: Color Enhance
          default: false
        original_name:
          type: string
          title: Original Name
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        created_at:
          type: string
          format: date-time
          title: Created At
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
        result_deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Result Deleted At
      type: object
      required:
      - id
      - status
      - scale
      - output_format
      - original_name
      - error
      - created_at
      - finished_at
      title: JobOut
...